+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 All notable changes will be documented here.
 
+## [1.0.1] - 15/01/2021
+### Added
+* GetAllPlaceholderData Endpoint
+### Changed
+* integration test now tests if deleted data still exists
+### Fixed
+
 ## [1.0.0] - 05/01/2021
 First release.